move generic pdf stuff from cal -> core
